Pivot On Default, Not Index
If there is a default class to the selected item, we can safely assume
that the user has not selected anything and that we should not highlight
the first object in the list. This also resolves the following user
story:

Given I am a user
When I click the first item in the dropdown
And I reclick the same dropdown
Then the dropdown highlights the first, selected item.